+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## 1.0.0-alpha.1 - 2021-09-20
-### Added
-- Initial release
+## 1.0.0-beta.4 - 2022-05-23
+
+### Changed
+- Removed encryption on employees of first name | last name | middle name | title
+- Removed encryption on employers of name